% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulm.R
\name{agg_predict}
\alias{agg_predict}
\title{Make model predictions aggregated by cell}
\usage{
agg_predict(
  object,
  ...,
  predX,
  predPsi,
  predTimes = NULL,
  prev_covar = NULL,
  counts = NULL,
  grouping_vars,
  alpha = 0.05,
  pop_df,
  K = NULL
)
}
\arguments{
\item{object}{ulm model output object}

\item{...}{other params}

\item{predX}{X matrix for the population}

\item{predPsi}{X matrix for the population}

\item{predTimes}{list of timepoints}

\item{prev_covar}{vector of previous responses, if known}

\item{counts}{number of units in each cell}

\item{alpha}{confidence level for CIs}

\item{K}{number of categories for categorical models}

\item{group_ids}{numeric ID to group and aggregate by}
}
\description{
Generates predictions for the population given a unit-level model
}
