% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur-package.R
\docType{package}
\name{blur-package}
\alias{blur}
\alias{blur-package}
\title{blur: Bayesian Longitudinal Unit-level Models in R}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel Vedensky \email{dvedensky@gmail.com}

Other contributors:
\itemize{
  \item Paul Parker \email{pparker1@ucsc.edu} [contributor]
  \item Scott Holan \email{holans@missouri.edu} [contributor]
}

}
\keyword{internal}
