% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulm.R
\name{coef.ulm}
\alias{coef.ulm}
\title{Get fixed effect coefficients (posterior means)}
\usage{
\method{coef}{ulm}(object, ..., include_reff = FALSE)
}
\arguments{
\item{object}{ulm model output object}

\item{...}{other params}

\item{include_reff}{logical value indicating whether to print random effect coefficients}
}
\description{
Get fixed effect coefficients (posterior means)
}
