% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{compute_mod_size}
\alias{compute_mod_size}
\title{Histogram of module sizes.}
\usage{
compute_mod_size(cones)
}
\arguments{
\item{cones}{martini results.}
}
\value{
Size of each SNP and median association score.
}
\description{
Plot an histogram with the size of the modules bigger than min_size.
}
