% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-binary-lon-mcmc.R
\name{fit_bin_lon_mcmc}
\alias{fit_bin_lon_mcmc}
\title{This function fits the PL-MB model via Gibbs sampling}
\usage{
fit_bin_lon_mcmc(
  Y,
  X,
  Psi,
  weights,
  timepoints,
  n_iter = 2000,
  n_burn = 500,
  init_vals = NULL,
  hyperparams = NULL
)
}
\arguments{
\item{Y}{A vector of binary survey responses}

\item{X}{The input covariate matrix, one row per sample unit}

\item{Psi}{The matrix of spatial basis functions, one row per sample unit}

\item{weights}{The vector of survey weights}

\item{timepoints}{list of timepoints}

\item{n_iter}{The number of iterations}

\item{n_burn}{The length of burn in}

\item{init_vals}{Initial values}

\item{hyperparams}{Fixed hyperparameters for fixed effects variance and inverse gamma priors}
}
\description{
This function fits the PL-MB model via Gibbs sampling
}
