% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-nominal-cs-mcmc.R
\name{fit_nominal_cs_mcmc}
\alias{fit_nominal_cs_mcmc}
\title{This function fits the nominal (unordered categorical) model via Gibbs sampling}
\usage{
fit_nominal_cs_mcmc(
  Y,
  X,
  Psi,
  n_iter,
  n_burn,
  weights = NULL,
  init_vals,
  hyperparams
)
}
\arguments{
\item{Y}{A one-hot matrix of survey responses}

\item{X}{The input covariate matrix, one row per sample unit}

\item{Psi}{The matrix of spatial basis functions, one row per sample unit}

\item{n_iter}{The number of iterations}

\item{n_burn}{The length of burn in}

\item{weights}{The vector of survey weights}

\item{init_vals}{Initial values}

\item{hyperparams}{Fixed hyperparameters for fixed effects variance and inverse gamma priors}
}
\description{
This function fits the nominal (unordered categorical) model via Gibbs sampling
}
