% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ordinal-cs-mcmc.R
\name{fit_ordinal_cs_mcmc}
\alias{fit_ordinal_cs_mcmc}
\title{Fit cross-sectional ordinal model with MCMC}
\usage{
fit_ordinal_cs_mcmc(
  Y,
  X,
  Psi,
  weights = NULL,
  n_iter = 2000,
  n_burn = 500,
  beta = NULL,
  init_vals = NULL,
  hyperparams = NULL
)
}
\arguments{
\item{Y}{n x K dimensional matrix of responses in one-hot format}

\item{X}{n x p design matrix of covariates}

\item{Psi}{n x r matrix of basis functions}

\item{weights}{n-dimensional vector of weights}

\item{n_iter}{number of iterations}

\item{n_burn}{number of iterations to discard for burn-in}

\item{init_vals}{Initial values}

\item{hyperparams}{Fixed hyperparameters for fixed effects variance and inverse gamma priors}
}
\value{
a list of matrices of MCMC chains for each parameter
}
\description{
Fit cross-sectional ordinal model with MCMC
}
