% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ordinal-cs-vb.R
\name{fit_ordinal_cs_vb}
\alias{fit_ordinal_cs_vb}
\title{This function fits the ordinal cross-sectional model via variational Bayes}
\usage{
fit_ordinal_cs_vb(
  Y,
  X,
  Psi,
  epsilon = 0.001,
  n_samples,
  weights = NULL,
  init_vals = NULL,
  hyperparams = NULL
)
}
\arguments{
\item{Y}{is a vector of binary survey responses}

\item{X}{is the input covariate matrix, one row per sample unit}

\item{Psi}{is the matrix of spatial basis functions, one row per sample unit}

\item{epsilon}{is the convergence tolerance}

\item{n_samples}{number of samples to draw from variational posterior}

\item{weights}{is the vector of survey weights}

\item{init_vals}{Initial values}

\item{hyperparams}{Fixed hyperparameters for fixed effects variance and inverse gamma priors}
}
\description{
This function fits the ordinal cross-sectional model via variational Bayes
}
