% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ordinal-lon-vb.R
\name{fit_ordinal_lon_vb}
\alias{fit_ordinal_lon_vb}
\title{This function fits the longitudinal ordinal model via variational Bayes}
\usage{
fit_ordinal_lon_vb(
  Y,
  X,
  Psi,
  weights,
  timepoints,
  epsilon = 0.001,
  n_samples,
  init_vals = NULL,
  hyperparams = NULL,
  prev_covar = NULL
)
}
\arguments{
\item{Y}{A vector of binary survey responses}

\item{X}{The input covariate matrix, one row per sample unit}

\item{Psi}{The matrix of spatial basis functions, one row per sample unit}

\item{weights}{The vector of survey weights}

\item{timepoints}{A vector of the time at which measurements were taken}

\item{epsilon}{is the convergence tolerance}

\item{n_samples}{number of samples to draw from variational posterior}

\item{init_vals}{Initial values}

\item{hyperparams}{Fixed hyperparameters for fixed effects variance and inverse gamma priors}

\item{prev_covar}{List of previous responses if longitudinal effect is desired}
}
\description{
This function fits the longitudinal ordinal model via variational Bayes
}
