% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirun.R
\name{join_experiments}
\alias{join_experiments}
\title{Join cones}
\usage{
join_experiments(labels, ...)
}
\arguments{
\item{labels}{Labels for the different experiments. Must be the same size of the number of experiments.}

\item{...}{Outputs of martini.}
}
\value{
Dataframe containing all the experiments.
}
\description{
Join the outputs of martini for different experiments
}
