% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{plot_size}
\alias{plot_size}
\title{Histogram of module sizes.}
\usage{
plot_size(cones, min_size = 1)
}
\arguments{
\item{cones}{martini results.}

\item{min_size}{Minimum size of the modules.}
}
\value{
An histogram of the module size.
}
\description{
Plot an histogram with the size of the modules bigger than min_size.
}
