% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{plot_size_association}
\alias{plot_size_association}
\title{Density plots of module sizes and the median association score of their SNPs.}
\usage{
plot_size_association(cones)
}
\arguments{
\item{cones}{martini results.}
}
\value{
An density plot of the module size and the median association score.
}
\description{
Density plots of module sizes and the median association score of their SNPs.
}
