% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{plot_subnetwork}
\alias{plot_subnetwork}
\title{Plot a subnetwork}
\usage{
plot_subnetwork(gwas, net, snps)
}
\arguments{
\item{gwas}{A GWAS experiment.}

\item{net}{A network of SNPs.}

\item{snps}{Logical indicating the rows of gwas$map that describe the subnetwork.}
}
\value{
A representation of the causal subnetwork.
}
\description{
Plot a subnetwork.
}
