% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulm.R
\name{predict.ulm}
\alias{predict.ulm}
\title{Make model predictions}
\usage{
\method{predict}{ulm}(object, ..., predX, predPsi, predTimes = NULL, counts = NULL, K = NULL)
}
\arguments{
\item{object}{ulm model output object}

\item{...}{other params}

\item{predX}{X matrix for the population}

\item{predPsi}{X matrix for the population}

\item{predTimes}{list of timepoints}

\item{counts}{number of units in each cell}

\item{K}{number of categories for categorical models}
}
\description{
Generates predictions for the population given a unit-level model
}
