% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Smooth.R
\name{smoothContinuousSubsets}
\alias{smoothContinuousSubsets}
\title{Smooth subset of continuous 3D data}
\usage{
smoothContinuousSubsets(r, coords, kernel, scales, edgeValues = c(),
  wide = T, .parallel = T)
}
\description{
Smooth subset of continuous 3D data
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}

