% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Smooth.R
\name{smoothDiscreteSubsets}
\alias{smoothDiscreteSubsets}
\title{Smooth subset of discrete 3D data}
\usage{
smoothDiscreteSubsets(r, coords, kernel, scales, processValues, edgeValues,
  wide = T, .parallel = T)
}
\description{
Smooth subset of discrete 3D data
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}

