% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulm.R
\name{summary.ulm}
\alias{summary.ulm}
\title{Summarize unit-level model fit}
\usage{
\method{summary}{ulm}(object, ..., include_reff = FALSE)
}
\arguments{
\item{object}{ulm model output object}

\item{...}{other params}

\item{include_reff}{indicate whether random effects should be included in summary output}
}
\description{
Summarize unit-level model fit
}
