% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulm.R
\name{ulm}
\alias{ulm}
\title{Constructor for unit level model}
\usage{
ulm(
  Y,
  X,
  Psi,
  weights = NULL,
  timepoints = NULL,
  n_iter = 2000,
  n_burn = 500,
  init_vals = NULL,
  hyperparams = NULL,
  longitudinal = FALSE,
  epsilon = NULL,
  n_samples = NULL,
  fix_sigma = NULL,
  prev_covar = NULL,
  response_type = "ordinal",
  algorithm = c("MCMC", "VB")
)
}
\arguments{
\item{Y}{vector or matrix of response values}

\item{X}{n x p design matrix of covariates}

\item{Psi}{n x r matrix of basis functions}

\item{weights}{n-dimensional vector of weights}

\item{timepoints}{list of timepoints (integer valued with first time point equal to 1)}

\item{n_iter}{number of iterations}

\item{n_burn}{number of iterations to discard for burn-in}

\item{init_vals}{list of initial values}

\item{hyperparams}{list of hyperparams}

\item{longitudinal}{logical value for whether to fit longitudinal model}

\item{epsilon}{tolerance for VB (VB only)}

\item{n_samples}{number of samples (VB only)}

\item{fix_sigma}{optional numeric value to fix sigma to in AL sampler (can be null)}

\item{prev_covar}{response values from the previous timepoint (long format)}

\item{response_type}{data model, one of "binary" or "ordinal"}

\item{algorithm}{sampling algorithm to employ, either "MCMC" or "VB"}
}
\value{
a list of matrices of MCMC chains for each parameter
}
\description{
Fits a Bayesian pseudolikelihood unit-level model for the given response type using the specified sampling algorithm
}
