% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirun.R
\name{consistency}
\alias{consistency}
\title{Consistency among different experiments}
\usage{
consistency(labels, ...)
}
\arguments{
\item{labels}{Labels for the different experiments. Must be the same size of the number of experiments.}

\item{...}{Outputs of martini.}
}
\value{
A heatmap with each of the SNPs selected in any of the experiments, and where they were selected.
}
\description{
Plot a heatmap displaying the consistency in the selection of different experiments.
}
