% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirun.R
\name{join_experiments}
\alias{join_experiments}
\title{Join cones}
\usage{
join_experiments(...)
}
\arguments{
\item{...}{Named outputs of martini e.g. '"Additive model" = cones_am, Recessive = cones_re'.}
}
\value{
Dataframe containing all the experiments.
}
\description{
Join the outputs of martini for different experiments
}
